Return back explicit creation of Linux kernel raw build commands file
Before it was read in process_all_linux_kernel_raw_build_cmds() with 1 second delay and make in build_linux_kernel() catches to create this file. But we can not rely on this always. Note that a long ago this file was created even early in configure_linux_kernel() but then this was disabled.