Make all relative paths absolute in final source files intended for verifiers
This allows to avoid so called paths normalization when working source tree was added for all relative file paths referred by error traces.

Besides it seems that this is required any way for reliable file referencing with merging (CIL) or/and verification since they will not be able to compute proper file paths if different CC commands will be executed in different directories.