VOG: Implement new Coverage aggregation strategy
The strategy is abstract and can be used anywhere but mainly intended
for verification of subsystems with modules. It allows to find for a
given target fragment that fragments that call export functions from
the target fragment in covered code. However, it requires to provide
an archive with coverage that can be downloaded from finished Jobs.