Get rid of unstable digital suffix in attribute values
This is necessary for CPALockator that specify declaration names as
attributes while these names may be artificial and contain unstable
parts, e.g. "struct __anonstruct_rdc321x_wdt_device_429".

Forthcoming update of CPALockator will likely fix the issue from the
other side, but it is not clear when this will happen.