Unconditionally remove former Core working directories if so
It is not clear why it is necessary to keep them if the user requests to
remove old stuff and to solve a job again.

BTW, this allowed to remove the pretty annoying check that always failed
(the existing job working directory occupies more disk space that is
allowed by the user). Moreover, it enables removing entirely option
"Ignore other instances of Klever Core" and related code in Core.