Specify actual list of supported Linux distributions
Also, point out actual links to download Python 3.10 binaries.

It is pretty hard to maintain more different Linux distributions as well
as other versions of supported Linux distributions because of various
incompatibility issues. For instance, Debian 9 has outdated version of
the openssl library (it is not enough for Python 3.10) and PostgreSQL
(new Django requires at least version 10). Moreover, it does not seem
that it is really necessary for anybody.