Fix tests and specifications to support changes in inodes structure
New versions of the Linux kernel introduces a new argument for some
inode callbacks ("struct user_namespace *" that is not modeled anyhow at
the moment).

Besides, field "i_bdev" was dropped. Current support completely removed
initialization of this field that can result in bad consequences
somewhere. In principle it is possible to define and use different
models or to initialize this extra field directly in code invoking these
models.