Copy more persistent sources and binaries earlier
They don't change often unlike .git, so Docker will be able to reuse
cached data rather than to spend a lot of time to copy them again and
again.

In addition, create stub top directories for generated build bases. This
is not necessary for complete build, but it may be necessary if one will
generate just some subsets of build bases.