Revert some changes in cross-ref collection
Cross-references has some dependencies in `Weaver`: cross-ref data is collected in `additional sources` directory with C files, which are used for coverage. And `additional sources` for every auxiliary file is used in cache. So, we cannot remove the directory in case of disabled cross-ref. Though, for now just disable `.idx` files, but potentially `additional sources` may be also optimized.

Also, fix a path of task description to plugin work directory.